requirements.txt
xvideos-api>=1.0.0
aiohttp>=3.8.0
Pillow>=9.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
